+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 1.2.0
+
+* Add fond settings for axis labels and data labels in cells
+* Cells scale depends on text size
+* Add auto-scroll if chart doesn't fit into viewport
+
+## 1.1.0
+
+* Add gradient color selection feature as alternative for pre defined colorbewers
+* Add bucket number limitation from 1 to 18 for gradient colors
+* Add bucket number limitation between min and max buckets in from set for pre defined colorbewers
+* X axis labels properties
+* Y axis labels properties
+* Add tolltips by tooltip services
+* Fix bug when left upper cell draws on axis label position if first value in the category is null
+* Left and top margins of grid depend from height and width values of axis labels
+* Add property to limit number of symbols in Y axis labels text
+
 ## 1.0.1
 
 * Data labels displaying in cells